chore: Move workflow file to correct directory
This commit moves the `action.yaml` workflow file from `.github/workflow/` to `.github/workflows/`. This change ensures that the workflow is correctly recognized and executed by GitHub Actions. The `repository` and `ref` properties in checkout action was simplified because there is no longer a need to checkout into a specific subdirectory.